Restore more core extensions
