+++ web/extensions/core/noteNode.js
@@ -1,7 +1,7 @@
 import { app } from "../../scripts/app.js";
 import { ComfyWidgets } from "../../scripts/widgets.js";
 import { ComfyGraphNode } from "../../scripts/graphNode.js";
-import { LiteGraph } from "../../lib/litegraph.core.js"
+import { LiteGraph, LGraphCanvas } from "../../lib/litegraph.core.js"
 
 // Node that add notes to your project
 
@@ -12,7 +12,8 @@ app.registerExtension({
             color=LGraphCanvas.node_colors.yellow.color;
             bgcolor=LGraphCanvas.node_colors.yellow.bgcolor;
             groupcolor = LGraphCanvas.node_colors.yellow.groupcolor;
-            constructor() {
+            constructor(title) {
+				super(title)
                 if (!this.properties) {
                     this.properties = {};
                     this.properties.text="";
@@ -30,11 +31,10 @@ app.registerExtension({
         LiteGraph.registerNodeType({
 			class: NoteNode,
             title_mode: LiteGraph.NORMAL_TITLE,
+			category: "utils",
 			type: "Note",
             title: "Note",
             collapsable: true,
 		});
-
-        NoteNode.category = "utils";
     },
 });
