+++ web/extensions/core/rerouteNode.js
@@ -8,7 +8,7 @@ app.registerExtension({
 	name: "Comfy.RerouteNode",
 	registerCustomNodes() {
 		class RerouteNode extends ComfyGraphNode {
-			constructor() {
+			onNodeCreated() {
 				if (!this.properties) {
 					this.properties = {};
 				}
@@ -18,137 +18,137 @@ app.registerExtension({
 				this.addInput("", "*");
 				this.addOutput(this.properties.showOutputText ? "*" : "", "*");
 
-				this.onConnectionsChange = function (type, index, connected, link_info) {
-					this.applyOrientation();
+				// This node is purely frontend and does not impact the resulting prompt so should not be serialized
+				this.isVirtualNode = true;
+			}
 
-					// Prevent multiple connections to different types when we have no input
-					if (connected && type === LiteGraph.OUTPUT) {
-						// Ignore wildcard nodes as these will be updated to real types
-						const types = new Set(this.outputs[0].links.map((l) => app.graph.links[l].type).filter((t) => t !== "*"));
-						if (types.size > 1) {
-							const linksToDisconnect = [];
-							for (let i = 0; i < this.outputs[0].links.length - 1; i++) {
-								const linkId = this.outputs[0].links[i];
-								const link = app.graph.links[linkId];
-								linksToDisconnect.push(link);
-							}
-							for (const link of linksToDisconnect) {
-								const node = app.graph.getNodeById(link.target_id);
-								node.disconnectInput(link.target_slot);
-							}
+			onConnectionsChange(type, index, connected, link_info) {
+				this.applyOrientation();
+
+				// Prevent multiple connections to different types when we have no input
+				if (connected && type === LiteGraph.OUTPUT) {
+					// Ignore wildcard nodes as these will be updated to real types
+					const types = new Set(this.outputs[0].links.map((l) => app.graph.links[l].type).filter((t) => t !== "*"));
+					if (types.size > 1) {
+						const linksToDisconnect = [];
+						for (let i = 0; i < this.outputs[0].links.length - 1; i++) {
+							const linkId = this.outputs[0].links[i];
+							const link = app.graph.links[linkId];
+							linksToDisconnect.push(link);
+						}
+						for (const link of linksToDisconnect) {
+							const node = app.graph.getNodeById(link.target_id);
+							node.disconnectInput(link.target_slot);
 						}
 					}
+				}
 
-					// Find root input
-					let currentNode = this;
-					let updateNodes = [];
-					let inputType = null;
-					let inputNode = null;
-					while (currentNode) {
-						updateNodes.unshift(currentNode);
-						const linkId = currentNode.inputs[0].link;
-						if (linkId !== null) {
-							const link = app.graph.links[linkId];
-							const node = app.graph.getNodeById(link.origin_id);
-							const type = node.constructor.type;
-							if (type === "Reroute") {
-								if (node === this) {
-									// We've found a circle
-									currentNode.disconnectInput(link.target_slot);
-									currentNode = null;
-								}
-								else {
-									// Move the previous node
-									currentNode = node;
-								}
-							} else {
-								// We've found the end
-								inputNode = currentNode;
-								inputType = node.outputs[link.origin_slot]?.type ?? null;
-								break;
+				// Find root input
+				let currentNode = this;
+				let updateNodes = [];
+				let inputType = null;
+				let inputNode = null;
+				while (currentNode) {
+					updateNodes.unshift(currentNode);
+					const linkId = currentNode.inputs[0].link;
+					if (linkId !== null) {
+						const link = app.graph.links[linkId];
+						const node = app.graph.getNodeById(link.origin_id);
+						const type = node.constructor.type;
+						if (type === "Reroute") {
+							if (node === this) {
+								// We've found a circle
+								currentNode.disconnectInput(link.target_slot);
+								currentNode = null;
+							}
+							else {
+								// Move the previous node
+								currentNode = node;
 							}
 						} else {
-							// This path has no input node
-							currentNode = null;
+							// We've found the end
+							inputNode = currentNode;
+							inputType = node.outputs[link.origin_slot]?.type ?? null;
 							break;
 						}
+					} else {
+						// This path has no input node
+						currentNode = null;
+						break;
 					}
+				}
 
-					// Find all outputs
-					const nodes = [this];
-					let outputType = null;
-					while (nodes.length) {
-						currentNode = nodes.pop();
-						const outputs = (currentNode.outputs ? currentNode.outputs[0].links : []) || [];
-						if (outputs.length) {
-							for (const linkId of outputs) {
-								const link = app.graph.links[linkId];
+				// Find all outputs
+				const nodes = [this];
+				let outputType = null;
+				while (nodes.length) {
+					currentNode = nodes.pop();
+					const outputs = (currentNode.outputs ? currentNode.outputs[0].links : []) || [];
+					if (outputs.length) {
+						for (const linkId of outputs) {
+							const link = app.graph.links[linkId];
 
-								// When disconnecting sometimes the link is still registered
-								if (!link) continue;
+							// When disconnecting sometimes the link is still registered
+							if (!link) continue;
 
-								const node = app.graph.getNodeById(link.target_id);
-								const type = node.constructor.type;
+							const node = app.graph.getNodeById(link.target_id);
+							const type = node.constructor.type;
 
-								if (type === "Reroute") {
-									// Follow reroute nodes
-									nodes.push(node);
-									updateNodes.push(node);
+							if (type === "Reroute") {
+								// Follow reroute nodes
+								nodes.push(node);
+								updateNodes.push(node);
+							} else {
+								// We've found an output
+								const nodeOutType = node.inputs && node.inputs[link?.target_slot] && node.inputs[link.target_slot].type ? node.inputs[link.target_slot].type : null;
+								if (inputType && nodeOutType !== inputType) {
+									// The output doesnt match our input so disconnect it
+									node.disconnectInput(link.target_slot);
 								} else {
-									// We've found an output
-									const nodeOutType = node.inputs && node.inputs[link?.target_slot] && node.inputs[link.target_slot].type ? node.inputs[link.target_slot].type : null;
-									if (inputType && nodeOutType !== inputType) {
-										// The output doesnt match our input so disconnect it
-										node.disconnectInput(link.target_slot);
-									} else {
-										outputType = nodeOutType;
-									}
+									outputType = nodeOutType;
 								}
 							}
-						} else {
-							// No more outputs for this path
 						}
+					} else {
+						// No more outputs for this path
 					}
+				}
 
-					const displayType = inputType || outputType || "*";
-					const color = app.canvas.link_type_colors[displayType];
+				const displayType = inputType || outputType || "*";
+				const color = app.canvas.link_type_colors[displayType];
 
-					// Update the types of each node
-					for (const node of updateNodes) {
-						// If we dont have an input type we are always wildcard but we'll show the output type
-						// This lets you change the output link to a different type and all nodes will update
-						node.outputs[0].type = inputType || "*";
-						node.__outputType = displayType;
-						node.outputs[0].name = node.properties.showOutputText ? displayType : "";
-						node.size = node.computeSize();
-						node.applyOrientation();
+				// Update the types of each node
+				for (const node of updateNodes) {
+					// If we dont have an input type we are always wildcard but we'll show the output type
+					// This lets you change the output link to a different type and all nodes will update
+					node.outputs[0].type = inputType || "*";
+					node.__outputType = displayType;
+					node.outputs[0].name = node.properties.showOutputText ? displayType : "";
+					node.size = node.computeSize();
+					node.applyOrientation();
 
-						for (const l of node.outputs[0].links || []) {
-							const link = app.graph.links[l];
-							if (link) {
-								link.color = color;
-							}
-						}
-					}
-
-					if (inputNode) {
-						const link = app.graph.links[inputNode.inputs[0].link];
+					for (const l of node.outputs[0].links || []) {
+						const link = app.graph.links[l];
 						if (link) {
 							link.color = color;
 						}
 					}
-				};
+				}
 
-				this.clone = function () {
-					const cloned = RerouteNode.prototype.clone.apply(this);
-					cloned.removeOutput(0);
-					cloned.addOutput(this.properties.showOutputText ? "*" : "", "*");
-					cloned.size = cloned.computeSize();
-					return cloned;
-				};
+				if (inputNode) {
+					const link = app.graph.links[inputNode.inputs[0].link];
+					if (link) {
+						link.color = color;
+					}
+				}
+			}
 
-				// This node is purely frontend and does not impact the resulting prompt so should not be serialized
-				this.isVirtualNode = true;
+			clone() {
+				const cloned = RerouteNode.prototype.clone.apply(this);
+				cloned.removeOutput(0);
+				cloned.addOutput(this.properties.showOutputText ? "*" : "", "*");
+				cloned.size = cloned.computeSize();
+				return cloned;
 			}
 
 			getExtraMenuOptions(_, options) {
@@ -224,11 +224,10 @@ app.registerExtension({
 		LiteGraph.registerNodeType({
 			class: RerouteNode,
 			title_mode: LiteGraph.NO_TITLE,
+			category: "utils",
 			type: "Reroute",
 			title: "Reroute",
 			collapsable: false,
 		});
-
-		RerouteNode.category = "utils";
 	},
 });
