+++ web/extensions/core/slotDefaults.js
@@ -28,7 +28,7 @@ app.registerExtension({
 	slot_types_default_out: {},
 	slot_types_default_in: {},
 	async beforeRegisterNodeDef(nodeType, nodeData, app) {
-				var nodeId = nodeData.name;
+		var nodeId = nodeData.name;
 		var inputs = [];
 		inputs = nodeData["input"]["required"]; //only show required inputs to reduce the mess also not logical to create node with optional inputs
 		for (const inputKey in inputs) {
@@ -79,7 +79,6 @@ app.registerExtension({
 		this.setDefaults(maxNum);
 	},
 	setDefaults(maxNum) {
-
 		LiteGraph.slot_types_default_out = {};
 		LiteGraph.slot_types_default_in = {};
