+++ web/scripts/app.js
@@ -245,369 +245,6 @@ export class ComfyApp {
 	}
 
 	/**
-	 * Adds special context menu handling for nodes
-	 * e.g. this adds Open Image functionality for nodes that show images
-	 * @param {*} node The node to add the menu handler
-	 */
-	#addNodeContextMenuHandler(node) {
-		node.prototype.getExtraMenuOptions = function (_, options) {
-			if (this.imgs) {
-				// If this node has images then we add an open in new tab item
-				let img;
-				if (this.imageIndex != null) {
-					// An image is selected so select that
-					img = this.imgs[this.imageIndex];
-				} else if (this.overIndex != null) {
-					// No image is selected but one is hovered
-					img = this.imgs[this.overIndex];
-				}
-				if (img) {
-					options.unshift(
-						{
-							content: "Open Image",
-							callback: () => {
-								let url = new URL(img.src);
-								url.searchParams.delete('preview');
-								window.open(url, "_blank")
-							},
-						},
-						{
-							content: "Save Image",
-							callback: () => {
-								const a = document.createElement("a");
-								let url = new URL(img.src);
-								url.searchParams.delete('preview');
-								a.href = url;
-								a.setAttribute("download", new URLSearchParams(url.search).get("filename"));
-								document.body.append(a);
-								a.click();
-								requestAnimationFrame(() => a.remove());
-							},
-						}
-					);
-				}
-			}
-
-			options.push({
-					content: "Bypass",
-					callback: (obj) => { if (this.mode === 4) this.mode = 0; else this.mode = 4; this.graph.change(); }
-				});
-
-			// prevent conflict of clipspace content
-			if(!ComfyApp.clipspace_return_node) {
-				options.push({
-						content: "Copy (Clipspace)",
-						callback: (obj) => { ComfyApp.copyToClipspace(this); }
-					});
-
-				if(ComfyApp.clipspace != null) {
-					options.push({
-							content: "Paste (Clipspace)",
-							callback: () => { ComfyApp.pasteFromClipspace(this); }
-						});
-				}
-
-				if(ComfyApp.isImageNode(this)) {
-					options.push({
-							content: "Open in MaskEditor",
-							callback: (obj) => {
-								ComfyApp.copyToClipspace(this);
-								ComfyApp.clipspace_return_node = this;
-								ComfyApp.open_maskeditor();
-							}
-						});
-				}
-			}
-		};
-	}
-
-	#addNodeKeyHandler(node) {
-		const app = this;
-		const origNodeOnKeyDown = node.prototype.onKeyDown;
-
-		node.prototype.onKeyDown = function(e) {
-			if (origNodeOnKeyDown && origNodeOnKeyDown.apply(this, e) === false) {
-				return false;
-			}
-
-			if (this.flags.collapsed || !this.imgs || this.imageIndex === null) {
-				return;
-			}
-
-			let handled = false;
-
-			if (e.key === "ArrowLeft" || e.key === "ArrowRight") {
-				if (e.key === "ArrowLeft") {
-					this.imageIndex -= 1;
-				} else if (e.key === "ArrowRight") {
-					this.imageIndex += 1;
-				}
-				this.imageIndex %= this.imgs.length;
-
-				if (this.imageIndex < 0) {
-					this.imageIndex = this.imgs.length + this.imageIndex;
-				}
-				handled = true;
-			} else if (e.key === "Escape") {
-				this.imageIndex = null;
-				handled = true;
-			}
-
-			if (handled === true) {
-				e.preventDefault();
-				e.stopImmediatePropagation();
-				return false;
-			}
-		}
-	}
-
-	/**
-	 * Adds Custom drawing logic for nodes
-	 * e.g. Draws images and handles thumbnail navigation on nodes that output images
-	 * @param {*} node The node to add the draw handler
-	 */
-	#addDrawBackgroundHandler(node) {
-		const app = this;
-
-		function getImageTop(node) {
-			let shiftY;
-			if (node.imageOffset != null) {
-				shiftY = node.imageOffset;
-			} else {
-				if (node.widgets?.length) {
-					const w = node.widgets[node.widgets.length - 1];
-					shiftY = w.last_y;
-					if (w.computeSize) {
-						shiftY += w.computeSize()[1] + 4;
-					}
-					else if(w.computedHeight) {
-						shiftY += w.computedHeight;
-					}
-					else {
-						shiftY += LiteGraph.NODE_WIDGET_HEIGHT + 4;
-					}
-				} else {
-					shiftY = node.computeSize()[1];
-				}
-			}
-			return shiftY;
-		}
-
-		node.prototype.setSizeForImage = function () {
-			if (this.inputHeight) {
-				this.setSize(this.size);
-				return;
-			}
-			const minHeight = getImageTop(this) + 220;
-			if (this.size[1] < minHeight) {
-				this.setSize([this.size[0], minHeight]);
-			}
-		};
-
-		node.prototype.onDrawBackground = function (ctx) {
-			if (!this.flags.collapsed) {
-				let imgURLs = []
-				let imagesChanged = false
-
-				const output = app.nodeOutputs[this.id + ""];
-				if (output && output.images) {
-					if (this.images !== output.images) {
-						this.images = output.images;
-						imagesChanged = true;
-						imgURLs = imgURLs.concat(output.images.map(params => {
-							return api.apiURL("/view?" + new URLSearchParams(params).toString() + app.getPreviewFormatParam());
-						}))
-					}
-				}
-
-				const preview = app.nodePreviewImages[this.id + ""]
-				if (this.preview !== preview) {
-					this.preview = preview
-					imagesChanged = true;
-					if (preview != null) {
-						imgURLs.push(preview);
-					}
-				}
-
-				if (imagesChanged) {
-					this.imageIndex = null;
-					if (imgURLs.length > 0) {
-						Promise.all(
-							imgURLs.map((src) => {
-								return new Promise((r) => {
-									const img = new Image();
-									img.onload = () => r(img);
-									img.onerror = () => r(null);
-									img.src = src
-								});
-							})
-						).then((imgs) => {
-							if ((!output || this.images === output.images) && (!preview || this.preview === preview)) {
-								this.imgs = imgs.filter(Boolean);
-								this.setSizeForImage?.();
-								app.graph.setDirtyCanvas(true);
-							}
-						});
-					}
-					else {
-						this.imgs = null;
-					}
-				}
-
-				if (this.imgs && this.imgs.length) {
-					const canvas = graph.list_of_graphcanvas[0];
-					const mouse = canvas.graph_mouse;
-					if (!canvas.pointer_is_down && this.pointerDown) {
-						if (mouse[0] === this.pointerDown.pos[0] && mouse[1] === this.pointerDown.pos[1]) {
-							this.imageIndex = this.pointerDown.index;
-						}
-						this.pointerDown = null;
-					}
-
-					let w = this.imgs[0].naturalWidth;
-					let h = this.imgs[0].naturalHeight;
-					let imageIndex = this.imageIndex;
-					const numImages = this.imgs.length;
-					if (numImages === 1 && !imageIndex) {
-						this.imageIndex = imageIndex = 0;
-					}
-
-					const shiftY = getImageTop(this);
-
-					let dw = this.size[0];
-					let dh = this.size[1];
-					dh -= shiftY;
-
-					if (imageIndex == null) {
-						let best = 0;
-						let cellWidth;
-						let cellHeight;
-						let cols = 0;
-						let shiftX = 0;
-						for (let c = 1; c <= numImages; c++) {
-							const rows = Math.ceil(numImages / c);
-							const cW = dw / c;
-							const cH = dh / rows;
-							const scaleX = cW / w;
-							const scaleY = cH / h;
-
-							const scale = Math.min(scaleX, scaleY, 1);
-							const imageW = w * scale;
-							const imageH = h * scale;
-							const area = imageW * imageH * numImages;
-
-							if (area > best) {
-								best = area;
-								cellWidth = imageW;
-								cellHeight = imageH;
-								cols = c;
-								shiftX = c * ((cW - imageW) / 2);
-							}
-						}
-
-						let anyHovered = false;
-						this.imageRects = [];
-						for (let i = 0; i < numImages; i++) {
-							const img = this.imgs[i];
-							const row = Math.floor(i / cols);
-							const col = i % cols;
-							const x = col * cellWidth + shiftX;
-							const y = row * cellHeight + shiftY;
-							if (!anyHovered) {
-								anyHovered = LiteGraph.isInsideRectangle(
-									mouse[0],
-									mouse[1],
-									x + this.pos[0],
-									y + this.pos[1],
-									cellWidth,
-									cellHeight
-								);
-								if (anyHovered) {
-									this.overIndex = i;
-									let value = 110;
-									if (canvas.pointer_is_down) {
-										if (!this.pointerDown || this.pointerDown.index !== i) {
-											this.pointerDown = { index: i, pos: [...mouse] };
-										}
-										value = 125;
-									}
-									ctx.filter = `contrast(${value}%) brightness(${value}%)`;
-									canvas.canvas.style.cursor = "pointer";
-								}
-							}
-							this.imageRects.push([x, y, cellWidth, cellHeight]);
-							ctx.drawImage(img, x, y, cellWidth, cellHeight);
-							ctx.filter = "none";
-						}
-
-						if (!anyHovered) {
-							this.pointerDown = null;
-							this.overIndex = null;
-						}
-					} else {
-						// Draw individual
-						const scaleX = dw / w;
-						const scaleY = dh / h;
-						const scale = Math.min(scaleX, scaleY, 1);
-
-						w *= scale;
-						h *= scale;
-
-						let x = (dw - w) / 2;
-						let y = (dh - h) / 2 + shiftY;
-						ctx.drawImage(this.imgs[imageIndex], x, y, w, h);
-
-						const drawButton = (x, y, sz, text) => {
-							const hovered = LiteGraph.isInsideRectangle(mouse[0], mouse[1], x + this.pos[0], y + this.pos[1], sz, sz);
-							let fill = "#333";
-							let textFill = "#fff";
-							let isClicking = false;
-							if (hovered) {
-								canvas.canvas.style.cursor = "pointer";
-								if (canvas.pointer_is_down) {
-									fill = "#1e90ff";
-									isClicking = true;
-								} else {
-									fill = "#eee";
-									textFill = "#000";
-								}
-							} else {
-								this.pointerWasDown = null;
-							}
-
-							ctx.fillStyle = fill;
-							ctx.beginPath();
-							ctx.roundRect(x, y, sz, sz, [4]);
-							ctx.fill();
-							ctx.fillStyle = textFill;
-							ctx.font = "12px Arial";
-							ctx.textAlign = "center";
-							ctx.fillText(text, x + 15, y + 20);
-
-							return isClicking;
-						};
-
-						if (numImages > 1) {
-							if (drawButton(x + w - 35, y + h - 35, 30, `${this.imageIndex + 1}/${numImages}`)) {
-								let i = this.imageIndex + 1 >= numImages ? 0 : this.imageIndex + 1;
-								if (!this.pointerDown || !this.pointerDown.index === i) {
-									this.pointerDown = { index: i, pos: [...mouse] };
-								}
-							}
-
-							if (drawButton(x + w - 35, y + 5, 30, `x`)) {
-								if (!this.pointerDown || !this.pointerDown.index === null) {
-									this.pointerDown = { index: null, pos: [...mouse] };
-								}
-							}
-						}
-					}
-				}
-			}
-		};
-	}
-
-	/**
 	 * Adds a handler allowing drag+drop of files onto the window to load workflows
 	 */
 	#addDropHandler() {
@@ -860,6 +497,9 @@ export class ComfyApp {
 		const defs = await api.getNodeDefs();
 		await this.registerNodesFromDefs(defs);
 		await this.#invokeExtensionsAsync("registerCustomNodes");
+
+		// Hide litegraph Subgraph from menu until it's implemented
+		LiteGraph.registered_node_types["graph/subgraph"].hide_in_node_lists = true;
 	}
 
     async registerNodesFromDefs(defs) {
@@ -882,7 +522,7 @@ export class ComfyApp {
 
             const ctor = class extends ComfyBackendNode {
                 constructor(title) {
-                    super(title, app, nodeId, nodeData);
+                    super(title, nodeId, nodeData);
                 }
             }
 
@@ -892,62 +532,7 @@ export class ComfyApp {
                 desc: `ComfyNode: ${nodeId}`
             }
 
-			// const node = Object.assign(
-			// 	function ComfyNode() {
-			// 		var inputs = nodeData["input"]["required"];
-			// 		if (nodeData["input"]["optional"] != undefined){
-			// 		    inputs = Object.assign({}, nodeData["input"]["required"], nodeData["input"]["optional"])
-			// 		}
-			// 		const config = { minWidth: 1, minHeight: 1 };
-			// 		for (const inputName in inputs) {
-			// 			const inputData = inputs[inputName];
-			// 			const type = inputData[0];
-
-			// 			if(inputData[1]?.forceInput) {
-			// 				this.addInput(inputName, type);
-			// 			} else {
-			// 				if (Array.isArray(type)) {
-			// 					// Enums
-			// 					Object.assign(config, widgets.COMBO(this, inputName, inputData, app) || {});
-			// 				} else if (`${type}:${inputName}` in widgets) {
-			// 					// Support custom widgets by Type:Name
-			// 					Object.assign(config, widgets[`${type}:${inputName}`](this, inputName, inputData, app) || {});
-			// 				} else if (type in widgets) {
-			// 					// Standard type widgets
-			// 					Object.assign(config, widgets[type](this, inputName, inputData, app) || {});
-			// 				} else {
-			// 					// Node connection inputs
-			// 					this.addInput(inputName, type);
-			// 				}
-			// 			}
-			// 		}
-
-			// 		for (const o in nodeData["output"]) {
-			// 			const output = nodeData["output"][o];
-			// 			const outputName = nodeData["output_name"][o] || output;
-			// 			const outputShape = nodeData["output_is_list"][o] ? LiteGraph.GRID_SHAPE : LiteGraph.CIRCLE_SHAPE ;
-			// 			this.addOutput(outputName, output, { shape: outputShape });
-			// 		}
-
-			// 		const s = this.computeSize();
-			// 		s[0] = Math.max(config.minWidth, s[0] * 1.5);
-			// 		s[1] = Math.max(config.minHeight, s[1]);
-			// 		this.size = s;
-			// 		this.serialize_widgets = true;
-
-			// 		app.#invokeExtensionsAsync("nodeCreated", this);
-			// 	},
-			// 	{
-			// 		title: nodeData.display_name || nodeData.name,
-			// 		comfyClass: nodeData.name,
-			// 	}
-			// );
-			// node.prototype.comfyClass = nodeData.name;
             node.type = nodeId;
-
-			// this.#addNodeContextMenuHandler(node);
-			// this.#addDrawBackgroundHandler(node, app);
-			// this.#addNodeKeyHandler(node);
 
 			await this.#invokeExtensionsAsync("beforeRegisterNodeDef", node, nodeData);
 			LiteGraph.registerNodeType(node);
